refactor(landing): extract PhoneMockup into its own file
